# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.51
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/teams:
    get:
      tags: [Teams]
      summary: List billing teams
      security:
        - ManagementSessionAuth: []
      parameters:
        - in: query
          name: includeArchived
          schema: { type: boolean, default: false }
      responses:
        "200": { description: Team list }
        "401": { description: Unauthorized }
    post:
      tags: [Teams]
      summary: Create a billing team with an optional soft shared budget
      security:
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamCreate"
      responses:
        "201": { description: Team created }
        "400": { description: Invalid request }
        "401": { description: Unauthorized }
        "409": { description: Team name already exists }
  /api/teams/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Teams]
      summary: Get a billing team and its active API-key assignments
      security:
        - ManagementSessionAuth: []
      responses:
        "200": { description: Team details }
        "401": { description: Unauthorized }
        "404": { description: Team not found }
    patch:
      tags: [Teams]
      summary: Update a billing team
      security:
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200": { description: Team updated }
        "400": { description: Invalid request }
        "401": { description: Unauthorized }
        "404": { description: Team not found }
        "409": { description: Team name already exists or the Team is archived }
    delete:
      tags: [Teams]
      summary: Archive a billing team and close its active key assignments
      security:
        - ManagementSessionAuth: []
      responses:
        "200": { description: Team archived }
        "401": { description: Unauthorized }
        "404": { description: Team not found }
  /api/teams/{id}/members:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Teams]
      summary: List active API-key billing assignments
      security:
        - ManagementSessionAuth: []
      responses:
        "200": { description: Active assignments }
        "401": { description: Unauthorized }
        "404": { description: Team not found }
    put:
      tags: [Teams]
      summary: Assign an API key to this billing team
      security:
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKeyId]
              properties:
                apiKeyId: { type: string, format: uuid }
      responses:
        "200": { description: Assignment created }
        "400": { description: Invalid request }
        "401": { description: Unauthorized }
        "404": { description: Team or API key not found }
        "409": { description: Assignment timestamp conflict }
    delete:
      tags: [Teams]
      summary: Close an API-key billing assignment
      security:
        - ManagementSessionAuth: []
      parameters:
        - in: query
          name: apiKeyId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Assignment closed }
        "400": { description: Invalid request }
        "401": { description: Unauthorized }
        "404": { description: Assignment not found }
  /api/teams/{id}/usage:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: startDate
        schema: { type: string, format: date-time }
      - in: query
        name: endDate
        schema: { type: string, format: date-time }
    get:
      tags: [Teams]
      summary: Get immutable Team usage attribution with per-key drill-down
      description: >-
        Reports estimated list cost separately from actual provider cost,
        subscription quota, and compression savings. Phase 1 soft budgets use
        committed usage and do not guarantee no overshoot under concurrency.
      security:
        - ManagementSessionAuth: []
      responses:
        "200": { description: Team usage report }
        "400": { description: Invalid range }
        "401": { description: Unauthorized }
        "404": { description: Team not found }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Two bearer families are accepted. Inference API keys (typically `sk-…`)
        authorize `/v1/*`. Management routes also accept `oma_live_…` Access Tokens
        (Settings → Access Tokens / `omniroute connect`) and API keys whose metadata
        includes `manage` or `admin` scope. See docs/guides/MANAGEMENT-AUTH.md.
        Bearer credentials are accepted on management routes that use this scheme;
        they are not rejected solely for being Bearer.
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie (auth_token) for protected management routes. Distinct from Bearer Access Tokens and API keys. See docs/guides/MANAGEMENT-AUTH.md.
...
  schemas:
    TeamCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        maxBudgetUsd: { type: number, exclusiveMinimum: 0, nullable: true }
        budgetDuration: { type: string, enum: [1d, 7d, 30d], nullable: true }
      description: maxBudgetUsd and budgetDuration must be configured or cleared together.
    TeamUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        maxBudgetUsd: { type: number, exclusiveMinimum: 0, nullable: true }
        budgetDuration: { type: string, enum: [1d, 7d, 30d], nullable: true }
      description: Send both budget fields to set a budget, or both as null to clear it.
...